Crate re_view_spatial

Rerun Spatial Views

Views that show entities in a 2D or 3D spatial relationship.

Modules

contexts ğŸ”’

eye ğŸ”’

heuristics ğŸ”’

max_image_dimension_subscriber ğŸ”’

mesh_cache ğŸ”’

mesh_loader ğŸ”’

pickable_textured_rect ğŸ”’

picking ğŸ”’

Handles picking in 2D and 3D spaces.

picking_ui ğŸ”’

picking_ui_pixel ğŸ”’

pinhole ğŸ”’

proc_mesh ğŸ”’

Procedurally-generated meshes for rendering objects that are specified geometrically, and have nontrivial numbers of vertices each, such as a sphere or cylinder.

scene_bounding_boxes ğŸ”’

space_camera_3d ğŸ”’

spatial_topology ğŸ”’

transform_cache ğŸ”’

ui ğŸ”’

ui_2d ğŸ”’

ui_3d ğŸ”’

view_2d ğŸ”’

view_2d_properties ğŸ”’

view_3d ğŸ”’

view_3d_properties ğŸ”’

view_kind ğŸ”’

visualizers ğŸ”’

Responsible for populating SceneSpatialPrimitives and SceneSpatialUiData

Structs

SpatialView2D

SpatialView3D

Functions

configure_background ğŸ”’

resolution_of_image_at ğŸ”’