Module re_view_spatial::contexts::transform_tree_context

Structs

TransformInfo

TransformTreeContext

Provides transforms from an entity to a chosen reference space for all elements in the scene for the currently selected time & timeline.

TransformsAtEntity 🔒

Resolved transforms at an entity.

TwoDInThreeDTransformInfo

Functions

debug_assert_transform_field_order 🔒

lookup_image_plane_distance 🔒

transform_from_pinhole_with_image_plane 🔒

transform_info_for_downward_propagation 🔒

Compute transform info for when we walk down the tree from the reference.

transform_info_for_upward_propagation 🔒

Compute transform info for when we walk up the tree from the reference.

transforms_at 🔒