re_view_spatial
Module picking_ui_pixel

Structs

PickedPixelInfo

TextureInteractionId

Identifies an image/texture interaction.

TextureReadbackUserdata 🔒

Enums

PixelValueSource 🔒

How we figure out what value to show for a single pixel.

Constants

ZOOMED_IMAGE_TEXEL_RADIUS 🔒

Functions

format_pixel_value 🔒

pixel_value_string_from_gpu_texture 🔒

pixel_value_string_from_image 🔒

pixel_value_ui 🔒

Shows the value of a pixel in an image. If no image info is provided, this only shows the position of the pixel.

show_zoomed_image_region

meter: iff this is a depth map, how long is one meter?

show_zoomed_image_region_area_outline 🔒

Draws a border for the area zoomed in by show_zoomed_image_region.

textured_rect_hover_ui

try_show_zoomed_image_region 🔒

meter: iff this is a depth map, how long is one meter?