re_view_spatial
Module transform_cache

Structs

CachedTransformsForTimeline

Cached transforms for a single timeline.

InvalidatedTransforms 🔒

Points in time that have changed for a given entity, i.e. the cache is invalid for these times.

ResolvedPinholeProjection

TransformAspect

Flags for the different kinds of independent transforms that the transform cache handles.

TransformCacheStoreSubscriber

Store subscriber that resolves all transform components at a given entity to an affine transform.

TransformsForEntity

Cached transforms for a single entity.

Functions

query_and_resolve_instance_poses_at_entity 🔒

Queries all components that are part of pose transforms, returning the transform from child to parent.

query_and_resolve_pinhole_projection_at_entity 🔒

query_and_resolve_tree_transform_at_entity 🔒

Queries all components that are part of pose transforms, returning the transform from child to parent.

query_view_coordinates

Queries view coordinates from either the archetypes::Pinhole or archetypes::ViewCoordinates archetype.

query_view_coordinates_at_closest_ancestor

Queries view coordinates from either the archetypes::Pinhole or archetypes::ViewCoordinates archetype at the closest ancestor of the given entity path.

Type Aliases

PinholeProjectionMap 🔒

Maps from time to pinhole projection.

PoseTransformMap 🔒