re_view_spatial::transform_cache
Type Alias PinholeProjectionMap

type PinholeProjectionMap = BTreeMap<TimeInt, Option<ResolvedPinholeProjection>>;

Maps from time to pinhole projection.

Unlike with tree & pose transforms, there’s no identity value that we can insert upon clears. (clears here meaning that the user first logs a pinhole and then later either logs a clear or an empty pinhole array) Therefore, we instead store those events as None values to ensure that everything after a clear is properly marked as having no pinhole projection.

Aliased Type

struct PinholeProjectionMap { /* private fields */ }