re_view_spatial::view_3d
Type Alias ViewType

type ViewType = Spatial3DView;

Aliased Type

struct ViewType {
    pub background: Background,
    pub line_grid: LineGrid3D,
    pub time_ranges: VisibleTimeRanges,
}

Fields
background: Background

Configuration for the background of the view.

line_grid: LineGrid3D

Configuration for the 3D line grid.

time_ranges: VisibleTimeRanges

Configures which range on each timeline is shown by this view (unless specified differently per entity).

If not specified, the default is to show the latest state of each component. If a timeline is specified more than once, the first entry will be used.