re_view_spatial::visualizers::utilities
Module entity_iterator

Functions

clamped_or

Iterate over all the values in the slice, then repeat the last value forever.

clamped_vec_or

Clamp the last value in values in order to reach a length of clamped_len.

clamped_vec_or_empty

Clamp the last value in values in order to reach a length of clamped_len.

iter_component

Iterate chunks as indexed deserialized batches.

iter_slices

Iterate chunks as indexed primitives.

process_archetype

Iterates through all entity views for a given archetype.