re_view_spatial::visualizers::utilities::entity_iterator
Function process_archetype

pub fn process_archetype<System: IdentifiedViewSystem, A, F>(
    ctx: &ViewContext<'_>,
    query: &ViewQuery<'_>,
    view_ctx: &ViewContextCollection,
    fun: F,
) -> Result<(), ViewSystemExecutionError>
where
    A: Archetype,
    F: FnMut(&QueryContext<'_>, &SpatialSceneEntityContext<'_>, &HybridResults<'_>) -> Result<(), ViewSystemExecutionError>,

Iterates through all entity views for a given archetype.

The callback passed in gets passed along a SpatialSceneEntityContext which contains various useful information about an entity in the context of the current scene.