re_view_spatial::visualizers::utilities::labels
Function show_labels_fallback

pub fn show_labels_fallback(
    ctx: &QueryContext<'_>,
    instance_count_component: &ComponentDescriptor,
    text_component: &ComponentDescriptor,
) -> ShowLabels

Given a visualizer’s query context, compute its ShowLabels fallback value (used when neither the logged data nor the blueprint provides a value).

Assumes that the visualizer reads the Text component for components. The instance_count_component parameter must be the component descriptor that defines the number of instances in the batch.

This function is normally used to implement the ComponentFallbackProvider that will be used in a LabeledBatch.