re_view_spatial::visualizers::videos
Function latest_at_query_video_from_datastore

fn latest_at_query_video_from_datastore(
    ctx: &ViewerContext<'_>,
    entity_path: &EntityPath,
) -> Option<(Arc<Result<Video, VideoLoadError>>, Blob)>

Queries a video from the datstore and caches it in the video cache.

Note that this does NOT check the blueprint store at all. For this, we’d need a re_viewer_context::DataResult instead of merely a EntityPath.

Returns None if there was no blob at the referenced path. Returns Some(Err(_)) if there was a blob but it failed to load for some reason. Errors are cached as well so loading a failed video won’t occur a high cost repeatedly.